Changemap and sort to use same basicReducer
