+++ src/redux/actions/simpleActions.js
@@ -1,0 +1,6 @@
+const setSelection = (prefix, selection) => ({
+  type: `${prefix}_SET_SELECTION`,
+  selection,
+});
+
+export default setSelection;
